src/main/java/io/zipcoder/crudapp/PersonService.java
package io.zipcoder.crudapp;


import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

@Service
public class PersonService {

    private PersonRepository repository;

    @Autowired
    public PersonService(PersonRepository repository) {
        this.repository = repository;
    }

    public Person create(Person person){
        return repository.save(person);
    }

    public void update(Person person) {
        repository.save(person);
    }
}
